requirements.txt
pandas>=1.3
pdfplumber>=0.7.6
python-dateutil>=2.8.2
pytest>=7.0
google-generativeai>=0.8.5
langgraph>=0.1.0
python-dotenv>=1.0.0